API update and some other minor cleanup/fixes
Main focus was to use alternative API that can incorporate the use of states and countries and that is also updated appropriately regarding information that previous API was not updating correctly. Also added a 5th Card to calculate predicted actual cases of COVID. Some other minor cleanups to imports not being used and to make the background and image color background uniform.